Modify Profile Image Setting

+++ Termproject/app/src/main/java/com/example/termproject/UserAdapter.kt
@@ -1,6 +1,5 @@
 package com.example.termproject
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
@@ -14,7 +13,7 @@ class UserAdapter(private val userList: List<MatchingRecyclingView.User>, privat
     inner class UserViewHolder(private val binding: ItemUserBinding) : RecyclerView.ViewHolder(binding.root) {
         fun bind(user: MatchingRecyclingView.User) {
             binding.ItemRank.text = getRank(user).toString()
-            binding.ItemProfile.setImageResource(user.profile.toInt())
+            binding.ItemProfile.setImageResource(user.profile)
             binding.ItemNick.text = user.nick
             binding.ItemScore.text = user.score.toString()
             when (user.status) {
